airflow-etl/requirements.txt
snowflake-connector-python
snowflake-sqlalchemy
pandas
pinecone-client
google-generativeai
openai
boto3
PyPDF2